+++ cmd/jwbcli/main_test.go
@@ -1,0 +1,5 @@
+package main
+
+func TestJwbcli() {
+	main()
+}
