--- main_test.go
@@ -1,7 +1,0 @@
-package main
-
-func ExampleHello() {
-	Hello()
-	// Output:
-	// hello, world!
-}
