+++ pkg/downloader/downloader.go
@@ -1,4 +1,4 @@
-package lib
+package downloader
 
 import (
 	"bytes"
