+++ pkg/vtt/vtt2txt.go
@@ -1,4 +1,4 @@
-package lib
+package vtt
 
 import (
 	"bufio"
